+++ src/utils/element.ts
@@ -1,29 +1,37 @@
 import { APP_REGEX } from '@/app-config'
 import { effectDispose, unrefElement, isValueType } from '@/utils'
 
-import type { PartialCSSStyleDeclaration, ElementSelector } from '@/types'
-import type { BasicTarget } from '@/types'
+import type {
+  BasicTarget,
+  QueryElementsOptions,
+  ElementSelector,
+} from '@/types'
 
 /**
  *
- * @param target Target element dom
- * @param className 所需添加className，可: 'xxx xxx' | 'xxx' 格式添加(参考向元素绑定 css 语法)
+ * @param target 目标元素或 ref 注册元素
+ * @param classNames 所需添加类名
  *
- * 添加元素className(可: 'xxx xxx' | 'xxx'格式添加)
+ * @description
+ * 为目标元素添加类名
  *
  * @example
- * targetDom 当前 class: a-class b-class
- * addClass(targetDom, 'c-class') => a-class b-class c-class
+ * // targetDom 当前 class: a-class b-class
+ * setClass(targetDom, 'c-class') // a-class b-class c-class
+ * setClass(targetDom, ['c-class', 'c-class']) // a-class b-class c-class
  */
-export const addClass = (
+export const setClass = (
   target: BasicTarget<Element | HTMLElement | SVGAElement>,
-  className: string,
+  classNames: string | string[],
 ) => {
   const update = () => {
     const element = unrefElement(target)
 
     if (element) {
-      const classes = className.trim().split(' ')
+      const classes =
+        typeof classNames === 'string'
+          ? classNames.trim().split(' ')
+          : classNames
 
       classes.forEach((item) => {
         if (item) {
@@ -42,30 +50,35 @@ export const addClass = (
 
 /**
  *
- * @param target Target element dom
- * @param className 所需删除className，可: 'xxx xxx' | 'xxx' 格式删除(参考向元素绑定 css 语法)
+ * @param target 目标元素或 ref 注册元素
+ * @param className 所需删除类名
  *
- * 删除元素className(可: 'xxx xxx' | 'xxx'格式删除)
- * 如果输入值为 removeAllClass 则会删除该元素所有 class name
+ * @description
+ * 为目标元素删除类名
  *
  * @example
- * targetDom 当前 class: a-class b-class
- * removeClass(targetDom, 'a-class') => b-class
+ * // targetDom 当前 class: a-class b-class
+ * removeClass(targetDom, 'a-class') // b-class
+ * removeClass(targetDom, ['a-class', 'b-class']) // null
+ * removeClass(targetDom, 'removeAllClass') // null
  */
 export const removeClass = (
   target: BasicTarget<Element | HTMLElement | SVGAElement>,
-  className: string | 'removeAllClass',
+  classNames: string | 'removeAllClass' | string[],
 ) => {
   const update = () => {
     const element = unrefElement(target)
 
     if (element) {
-      if (className === 'removeAllClass') {
+      if (classNames === 'removeAllClass') {
         const classList = element.classList
 
         classList.forEach((curr) => classList.remove(curr))
       } else {
-        const classes = className.trim().split(' ')
+        const classes =
+          typeof classNames === 'string'
+            ? classNames.trim().split(' ')
+            : classNames
 
         classes.forEach((item) => {
           if (item) {
@@ -85,15 +98,20 @@ export const removeClass = (
 
 /**
  *
- * @param target Target element dom
- * @param className 查询元素是否含有此className，可: 'xxx xxx' | 'xxx' 格式查询(参考向元素绑定 css 语法)
+ * @param target 目标元素或 ref 注册元素
+ * @param className 查询元素是否含有此类名
  *
- * 元素是否含有某个className(可: 'xxx xxx' | 'xxx' 格式查询)
+ * @description
+ * 查询元素是否含有此类名
  *
  * @example
- * hasClass(targetDom, 'matchClassName') => Ref<true> | Ref<false>
+ * hasClass(targetDom, 'matchClassName') // Ref<true> | Ref<false>
+ * hasClass(targetDom, ['matchClassName', 'matchClassName']) // Ref<true> | Ref<false>
  */
-export const hasClass = (target: BasicTarget<Element>, className: string) => {
+export const hasClass = (
+  target: BasicTarget<Element>,
+  classNames: string | string[],
+) => {
   const hasClassRef = ref(false)
 
   const update = () => {
@@ -104,12 +122,17 @@ export const hasClass = (target: BasicTarget<Element>, className: string) => {
     } else {
       const elementClassName = element.className
 
-      const classes = className
-        .trim()
-        .split(' ')
-        .filter((item: string) => item !== '')
+      const classes =
+        typeof classNames === 'string'
+          ? classNames
+              .trim()
+              .split(' ')
+              .filter((curr: string) => curr !== '')
+          : classNames
 
-      hasClassRef.value = elementClassName.includes(classes.join(' '))
+      hasClassRef.value = classes.some((curr) =>
+        elementClassName.includes(curr),
+      )
     }
   }
 
@@ -124,15 +147,43 @@ export const hasClass = (target: BasicTarget<Element>, className: string) => {
 
 /**
  *
+ * @param style 所需添加样式
+ *
+ * @returns 添加前缀后的样式
+ *
+ * @description
+ * 为样式添加浏览器前缀，返回一个对象
+ *
+ * @example
+ * autoPrefixStyle('transform') => {webkitTransform: 'transform', mozTransform: 'transform', msTransform: 'transform', oTransform: 'transform'}
+ */
+export const autoPrefixStyle = (style: string) => {
+  const prefixes = ['webkit', 'moz', 'ms', 'o']
+  const styleWithPrefixes = {}
+
+  prefixes.forEach((prefix) => {
+    styleWithPrefixes[
+      `${prefix}${style.charAt(0).toUpperCase()}${style.slice(1)}`
+    ] = style
+  })
+
+  return styleWithPrefixes
+}
+
+/**
+ *
  * @param target Target element dom
  * @param styles 所需绑定样式(如果为字符串, 则必须以分号结尾每个行内样式描述)
+ *
+ * @description
+ * 为目标元素添加样式
  *
  * @example
  * style of string
  * ```
  * const styles = 'width: 100px; height: 100px; background: red;'
  *
- * addStyle(styles)
+ * setStyle(styles)
  * ```
  * style of object
  * ```
@@ -141,14 +192,37 @@ export const hasClass = (target: BasicTarget<Element>, className: string) => {
  *   height: '100px',
  * }
  *
- * addStyle(styles)
+ * setStyle(styles)
  * ```
  */
-export const addStyle = (
+export const setStyle = (
   target: BasicTarget<HTMLElement | SVGAElement>,
-  styles: PartialCSSStyleDeclaration | string,
+  styles: Partial<CSSStyleDeclaration> | string | string[],
 ) => {
-  let styleObj: PartialCSSStyleDeclaration
+  const set = (styleStr: string, element: HTMLElement | SVGAElement) => {
+    styleStr.split(';').forEach((curr) => {
+      const [key, value] = curr.split(':')
+
+      if (key && value) {
+        const trimKey = key.trim()
+        const trimValue = value.trim()
+
+        // 是否为 css variable
+        if (key.startsWith('--')) {
+          element.style.setProperty(trimKey, trimValue)
+        } else {
+          // 兼容浏览器前缀
+          const kitFix = autoPrefixStyle(trimKey)
+
+          Object.keys(kitFix).forEach((key) => {
+            element.style[key] = kitFix[key]
+          })
+          // 设置默认需要添加样式
+          element.style[trimKey] = trimValue
+        }
+      }
+    })
+  }
 
   const update = () => {
     const element = unrefElement(target)
@@ -158,26 +232,18 @@ export const addStyle = (
     }
 
     if (isValueType<string>(styles, 'String')) {
-      styleObj = styles.split(';').reduce((pre, curr) => {
-        const [key, value] = curr.split(':').map((s) => s.trim())
-
-        if (key && value) {
-          pre[key] = value
-        }
-
-        return pre
-      }, {} as PartialCSSStyleDeclaration)
+      set(styles, element)
+    } else if (isValueType<string[]>(styles, 'Array')) {
+      styles.forEach((curr) => {
+        set(curr, element)
+      })
     } else {
-      styleObj = styles
+      const keys = Object.keys(styles)
+
+      keys.forEach((curr) => {
+        set(`${curr}: ${styles[curr]}`, element)
+      })
     }
-
-    Object.keys(styleObj).forEach((key) => {
-      const value = styleObj[key]
-
-      if (key in element!.style) {
-        element!.style[key] = value
-      }
-    })
   }
 
   const watcher = watch(() => unrefElement(target), update, {
@@ -191,6 +257,9 @@ export const addStyle = (
  *
  * @param el Target element dom
  * @param styles 所需卸载样式
+ *
+ * @description
+ * 为目标元素卸载样式
  *
  * 当你发现不能正常的移除某些样式的时候，应该考虑是否是样式表兼容问题
  *
@@ -225,14 +294,15 @@ export const removeStyle = (
  * @param color 颜色格式
  * @param alpha 透明度
  *
+ * @description
  * 将任意颜色值转为 rgba，如果本身为 rgba, rgb 或者其它非法颜色值则直接返回
  *
  * @example
- * colorToRgba('#123632', 0.8) => rgba(18, 54, 50, 0.8)
- * colorToRgba('rgb(18, 54, 50)', 0.8) => rgb(18, 54, 50)
- * colorToRgba('#ee4f12', 0.3) => rgba(238, 79, 18, 0.3)
- * colorToRgba('rgba(238, 79, 18, 0.3)', 0.3) => rgba(238, 79, 18, 0.3)
- * colorToRgba('not a color', 0.3) => not a color
+ * colorToRgba('#123632', 0.8) // rgba(18, 54, 50, 0.8)
+ * colorToRgba('rgb(18, 54, 50)', 0.8) // rgb(18, 54, 50)
+ * colorToRgba('#ee4f12', 0.3) // rgba(238, 79, 18, 0.3)
+ * colorToRgba('rgba(238, 79, 18, 0.3)', 0.3) // rgba(238, 79, 18, 0.3)
+ * colorToRgba('not a color', 0.3) // not a color
  */
 export const colorToRgba = (color: string, alpha = 1) => {
   const hexPattern = /^#([0-9a-f]{3}|[0-9a-f]{6}|[0-9a-f]{8})$/i
@@ -271,24 +341,31 @@ export const colorToRgba = (color: string, alpha = 1) => {
  * @param element 需要匹配元素参数名称
  * @returns 匹配元素列表
  *
- * @remark 使用 querySelectorAll 作为检索方法
- * @remark 如果希望按照 attribute 匹配, 仅需要 'attr:xxx'传递参数即可
+ * @description
+ * 使用 querySelectorAll 作为检索方法。
+ *
+ * 如果希望按照 attribute 匹配, 仅需要 'attr:xxx'传递参数即可
  *
  * @example
- * class:
+ * // class:
  * const el = queryElements('.demo')
- * id:
+ * // id:
  * const el = queryElements('#demo')
- * attribute:
+ * // attribute:
  * const el = queryElements('attr:type=button')
- * 或者可以这样写
+ * // 或者可以这样写
  * const el = queryElements('attr:type')
+ * // 默认元素
+ * const el = queryElements('.demo', { defaultElement: document.body })
  */
 export const queryElements = <T extends Element = Element>(
   selector: ElementSelector,
+  options?: QueryElementsOptions<T>,
 ) => {
+  const { defaultElement } = options || {}
+
   if (!selector) {
-    return null
+    return defaultElement ? [defaultElement] : null
   }
 
   const queryParam = selector.startsWith('attr:')
@@ -300,9 +377,12 @@ export const queryElements = <T extends Element = Element>(
 
     return elements
   } catch (error) {
-    console.error(`Failed to get elements for selector '${selector}'`, error)
+    console.error(
+      `[queryElements]: Failed to get elements for selector '${selector}'`,
+      error,
+    )
 
-    return null
+    return defaultElement ? [defaultElement] : null
   }
 }
 
@@ -311,7 +391,8 @@ export const queryElements = <T extends Element = Element>(
  * @param size css size
  * @param unit 自动填充 css 尺寸单位
  *
- * @remark 自动补全尺寸
+ * @description
+ * 自动补全尺寸
  */
 export const completeSize = (size: number | string, unit = 'px') => {
   if (typeof size === 'number') {